pyproject.toml
[tool.poetry]
name = "reportedai"
version = "0.1.0"
description = ""
authors = ["mourtallah <serigne.faye@icloud.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.8"
psycopg = "^3.0.15"
psycopg-binary = "^3.0.15"
Django = "^4.0.6"
djangorestframework = "^3.13.1"
psycopg2 = "^2.9.3"
djangorestframework-simplejwt = "^5.2.0"
django-rest-passwordreset = "^1.2.1"
django-cors-headers = "^3.13.0"
django-import-export = "^2.8.0"
requests = "^2.28.1"
confluent-kafka = "^1.9.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
